+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## Pre-release Versions
 
+### v0.4.0
+
+- 
+
 ### v0.3.0
 
 - **[Update]** Internal improvements.